refactor: split response formatting out of McpResponse
