+++ AGENTS.md
@@ -17,3 +17,17 @@ This repository contains an MCP server and CLI for Chrome DevTools.
 - Do not use `// @ts-nocheck` comments.
 - Do not use `// @ts-expect-error` comments.
 - Prefer `for..of` instead of `forEach`.
+
+## Testing Layering
+
+When writing tests, follow this layered approach:
+
+1. **Tool Tests (`tests/tools/*.test.ts`)**: Should test that tools correctly
+   configure the `McpResponse` based on their inputs.
+2. **Data Fetching Tests (`tests/McpResponse.test.ts`)**: Should test that data
+   is correctly fetched from the browser based on the `McpResponse`
+   configuration. These tests should use snapshots. Data fetching should happen
+   in McpResponse ONLY if it is used by more than one tool.
+3. **Formatter Tests (`tests/formatters/*.test.ts`)**: Should test that mock
+   data is formatted and output correctly by the formatters. These tests should
+   use snapshots.
