+++ tests/tools/input.test.ts
@@ -27,7 +27,7 @@ import {
 } from '../../src/tools/input.js';
 import {parseKey} from '../../src/utils/keyboard.js';
 import {serverHooks} from '../server.js';
-import {html, withMcpContext, getTextContent} from '../utils.js';
+import {html, withMcpContext, testHandle, getTextContent} from '../utils.js';
 
 describe('input', () => {
   const server = serverHooks();
@@ -155,7 +155,7 @@ describe('input', () => {
           response,
           context,
         );
-        const result = await response.handle('click', context);
+        const result = await testHandle(response, 'click', context);
         const textContent = getTextContent(result.content[0]);
         const expectedUrl = server.getRoute('/after-click');
         assert.ok(
@@ -184,7 +184,7 @@ describe('input', () => {
           response,
           context,
         );
-        const result = await response.handle('click', context);
+        const result = await testHandle(response, 'click', context);
         const textContent = getTextContent(result.content[0]);
         assert.ok(
           !textContent.includes('Page navigated to '),
